requirements.txt
django
pillow
django-crispy-forms
cloudinary
